@@ -23,7 +23,7 @@
 		"postcss": "^8.4.35",
 		"prettier": "3.2.5",
 		"prettier-plugin-svelte": "^3.2.2",
-		"svelte": "^4.2.7",
+		"svelte": "^5.0.0",
 		"svelte-check": "^3.6.0",
 		"tailwindcss": "^3.4.1",
 		"tslib": "^2.4.1",
@@ -32,7 +32,6 @@
 	},
 	"type": "module",
 	"dependencies": {
-		"@jill64/svelte-sanitize": "^1.2.12",
 		"@sveltejs/svelte-virtual-list": "^3.0.1",
 		"@tauri-apps/api": "^2.0.0-beta.15",
 		"@tauri-apps/plugin-cli": "^2.0.0-beta.7",
@@ -46,7 +45,7 @@
 		"@tauri-apps/plugin-process": "^2.0.0-beta.7",
 		"@tauri-apps/plugin-shell": "^2.0.0-beta.8",
 		"@tauri-apps/plugin-updater": "^2.0.0-beta.7",
-		"bits-ui": "0.19.6",
+		"bits-ui": "0.21.13",
 		"rehype-highlight": "^7.0.0",
 		"rehype-pretty-code": "^0.14.0",
 		"rehype-raw": "^7.0.0",
