@@ -13,8 +13,8 @@ use tauri::{
 
 mod actions;
 mod commands;
+mod get;
 pub mod install;
-mod query;
 
 pub fn init() -> TauriPlugin<tauri::Wry> {
     Builder::new("gale-profile")
@@ -25,13 +25,24 @@ pub fn init() -> TauriPlugin<tauri::Wry> {
             let handle = app.app_handle().to_owned();
             async_runtime::spawn(install::handler(handle));
 
+            let handle = app.app_handle().to_owned();
+            async_runtime::spawn(async move {
+                let state = handle.app_state();
+                let path = PathBuf::from(r"D:\Gale\v2\Default");
+
+                match actions::create("Default", &path, 1, state).await {
+                    Ok(id) => log::info!("created default profile with id: {}", id),
+                    Err(err) => log::error!("failed to create default profile: {:#}", err),
+                };
+            });
+
             Ok(())
         })
         .invoke_handler(generate_handler![
             commands::create,
             commands::delete,
             commands::rename,
-            commands::query,
+            commands::get,
             commands::force_uninstall_mod,
             commands::force_toggle_mod,
             commands::queue_thunderstore_install,
@@ -51,7 +62,7 @@ enum ProfileModSource {
 }
 
 async fn emit_update(id: i64, state: &AppState, app: &AppHandle) -> Result<()> {
-    let info = query::single(id, state).await?;
+    let info = get::single(id, state).await?;
     app.emit("profile-update", info)?;
     Ok(())
 }
